requirements_2.x.txt
Beaker==1.8.0
funcsigs==1.0.2
genedataset==0.6
itsdangerous==0.24
Mako==1.0.4
MarkupSafe==0.23
nose==1.3.7
numexpr==2.6.1
numpy==1.11.1
pandas==0.18.1
PasteDeploy==1.5.2
Pygments==2.1.3
pyramid==1.5.7
pyramid-beaker==0.8
pyramid-debugtoolbar==3.0.2
pyramid-mailer==0.14.1
pyramid-mako==1.0.2
pyramid-tm==0.12.1
python-dateutil==2.5.3
pytz==2016.6.1
repoze.lru==0.6
repoze.sendmail==4.2
rpy2==2.8.2
selenium==2.53.6
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.1.0b2
tables==3.2.3.1
transaction==1.6.1
translationstring==1.3
venusian==1.0
waitress==1.0a2
WebOb==1.4.1
zope.deprecation==4.1.2
zope.interface==4.1.2
zope.sqlalchemy==0.7.7
